+++ internal/archive/archive.go
@@ -1,5 +1,9 @@
 package archive
 
+import (
+	"regexp"
+)
+
 const (
 	// ItemInfoStatusNew refers to an item that appeared for the first time in the
 	// archive.
@@ -91,7 +95,7 @@ func (a Info) FilterByStatuses(statuses ...ItemInfoStatus) Info {
 // Archive manages an archive joining all paths in a file, extracting and
 // calculating Checksums.
 type Archive interface {
-	Build(lastArchiveInfo Info, backupPaths ...string) (string, Info, error)
+	Build(lastArchiveInfo Info, ignoreFiles *regexp.Regexp, backupPaths ...string) (string, Info, error)
 	Extract(filename string, filter []string) (Info, error)
 	FileChecksum(filename string) (string, error)
 }